greatly changed project structure
