+++ app/handlers/text_handler.py
@@ -2,18 +2,10 @@
 from telebot import TeleBot
 from telebot.types import Message
 
-# for generating random strings
+from app.services.text_service import generate_random_text
 import string
-import random
 
 SPECIAL = ['_', '*', '[', ']', '(', ')', '~', '`', '<' , '>', '#', '+', '-', '=', '|', '{', '}', '.', '!' ]
-
-#  random text generater for the following handlers
-def generate_random_text():
-    length = random.randint(30,120)
-    return ''.join(random.choices(string.ascii_letters +
-                                  string.digits +
-                                  string.punctuation, k=length))
 
 # invoke with "/random" command
 def get_random_text(message: Message, bot: TeleBot):
